Test out the rest of the config parser

+++ bolo.h
@@ -99,7 +99,7 @@ void debugf(const char *fmt, ...);
 struct config {
 	int log_level;
 	char *secret_key;
-	unsigned int block_span;
+	unsigned long block_span;
 };
 
 int configure(struct config *, int fd) RETURNS;
